refactor: basic simulator
